app/parsers/windows_events/evtxtoelk/requirements.txt
python-evtx
elasticsearch
xmltodict
